Moving writes into redshift.go.
